Set up CouchDB and checks that a match is only added once.
Right now I have a console application to test the CouchDB and leaguelib
functionality of the LoLStatTracker. Not all of the Java libraries
needed to run this project are included in the lib folder, that will be
in a future commit. A new database is created for each summoner, but the
design document needed to use the database is not included yet. In the
future, when performing any DB queries if a 'Document Not Found'
exception occurs then a local copy of the design document will be posted
to the database. Still a huge work in progress.