Item list and match json
I added a LeagueItem class based heavily on the LeagueItem class from
the LeagueLibExtended repository. It includes the name, cost, and item
id of all the items. Additionally this commit includes the match json
document that is used to setup new summoners with couchdb. It is the
design document for matches and it is automatically added to a couchdb
table if the DocumentNotFound exception occurs