Change main summoner, some error handling
I added the ability to change main summoner. Additionally, I added some
exception handling for the login process. If the user enters the wrong
username/password/server they will be prompted to do it again. If the
server version is incorrect, the user will be prompted to change that.
Most of the println commands used for debugging have also been removed.