Updated build script
I changed the build script from the LeagueLib build to a build script
for my own use. It is completely different because it is being used now
to build LoLStatTracker. I included some of the resources that also go
into making the executable. Additionally, I included the missing Twin
Shadows icon